add header rename for fuzz move
