+++ programs/fuzz/fuzz_common.h
